Updated to Jorgen-Viking God and Adafruit GFX, using LEDtest ArronLdament for more customizable quick text. Some patterns still need testing, notably the scroll Text needs considerable updating
