--- LED2560_8x32S/macros.h
@@ -1,15 +1,0 @@
-// Varidaic Debug Macro
- // #define DEBUG   //Comment this line to disable Debug output
-#ifdef DEBUG    // Debug is on
-#define DBEGIN(...)    Serial.begin(__VA_ARGS__)
-#define DPRINT(...)    Serial.print(__VA_ARGS__)     //Sends our arguments to DPRINT()
-#define DPRINTLN(...)  Serial.println(__VA_ARGS__)   //Sends our arguments to DPRINTLN()
-#define DFLUSH(...)    Serial.flush()
-
-#else // Debug is off
-#define DBEGIN(...)
-#define DPRINT(...)     //Nothing Happens
-#define DPRINTLN(...)   //Nothing Happens
-#define DFLUSH(...)
-
-#endif // end macro
