--- LED2560_8x32S/messages.h
@@ -1,11 +1,0 @@
-// Scrolling messages
-
-const char string0[] PROGMEM = "TESTING TEXT STRINGS";
-const char string1[] PROGMEM = "autoNoiseCreationStation is Online  ";
-const char string2[] PROGMEM = "www.autoAV.co.uk  ";
-
-const char * const stringArray[] PROGMEM = {
-  string0,
-  string1,
-  string2
-};
