+++ LED2560_8x8T_R2/macros.h
@@ -1,5 +1,5 @@
 // Varidaic Debug Macro
-  #define DEBUG   //Comment this line to disable Debug output
+  //#define DEBUG   //Comment this line to disable Debug output
 #ifdef DEBUG    // Debug is on
 #define DBEGIN(...)    Serial.begin(__VA_ARGS__)
 #define DPRINT(...)    Serial.print(__VA_ARGS__)     //Sends our arguments to DPRINT()
